feat(deposit-slice): move to parent folder
