+++ src/redux/store.ts
@@ -5,7 +5,7 @@ import deviceHealthReducer from './deviceHealthCheck/slice'
 import pinnedMessageReducer from './PinnedMessage/slice'
 import execClientReducer from './ValidatorOnboarding/ClientSetup/slice'
 import keyGenerationReducer from './ValidatorOnboarding/KeyGeneration/slice'
-import depositReducer from './ValidatorOnboarding/Deposit/slice'
+import depositReducer from './Deposit/slice'
 import leftSidebarReducer from './LeftSidebar/slice'
 import rightSidebarReducer from './RightSidebar/slice'
 import validatorOnboardingReducer from './ValidatorOnboarding/slice'
